entity-update

+++ src/main/java/com/spring3/oauth/jwt/repositories/UserRepository.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface UserRepository extends RefreshableCRUDRepository<User, Long> {
+public interface UserRepository extends RefreshableCRUDRepository<User, Integer> {
 
    public User findByUsername(String username);
-   User findFirstById(Long id);
+   User findFirstById(Integer id);
 
 }
